requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyyaml==6.0.2
redis==5.0.7
pydantic==2.9.2
jinja2==3.1.4
python-multipart==0.0.9
orjson==3.10.7
pytest==8.3.3
